Больше ООП для loyalty
