+++ lib/loyalty_modules/__init__.py
@@ -1,0 +1,1 @@
+from .check_status import Check_Status
